backend: refactor corpse logic
